Update and rename MultiAucSim to MultiAucSim.py
 - Updated filename to include Python extension.

# Don't fully understand the purpose but I thought I would try and clean up some stuff you have going on.
 - Added a Player class instead of using Strings to represent players. This will allow us to assign Players properties that we can extend later. More importantly, also adding an ID field, (Can use Razzballs later after export/scrapping or webapi call)

Line 22: I would change this to a list of owner objects, rather than strings or a set of strings.
Line 24-26: I changed our value representation of player values to reference said ID from above. Note: maybe add default value for player in case user does not specify a value.
Line 42: Added a buildUniverse function that would house the retreival/building of player Objects.

Did not touch the bidwar function too much because I do not fully understand the purpose.
But I do know Line 57 & 60 & 63 & 98 &100 &102 were using 'p'. p was not in our current scope and since Python is functional it will actually look for variables outside it's current scope and back to it's parent so that is why there was no error. However this will most definitely lead to disaster, so at the very least out of all my changes I would use the parameter you are already passing in labeled 'player'.